refactor(acp): Extract tests to integration files, strip narrating comments
Move test blocks from 5 mod.rs files into dedicated integration test
files (112 tests total). Strip "what" comments that narrate code from
document_acp.rs, memory.rs, persistent.rs. Remove "Tests extracted"
change history comments from persistent.rs and local.rs.
